requirements.txt
streamlit
streamlit-audiorec
google-generativeai
